Remove single-transaction flag in script
Running a derived table query file in a transaction causes the drop
table statement to be rolled back if an error occurs while creating
the new table.  Allowing the drop table to be committed in such cases
is preferable to the possibility that users might query an old version
of the table without realizing it.  For this reason the example script
has been changed to run the SQL in autocommit mode.