Convert language subquery to join (incomplete)
The subquery selecting language from
folio_reporting.instance_languages causes the query to run for hours,
when testing with a large database.  Converting the subquery to a join
addresses this problem.  This conversion is not yet complete, because
there may be duplicate rows which previously had been removed with
"LIMIT 1" in the subquery.