Update shelf_list.sql
Added comments on the hard coded filters for item, holdings, and instance discovery suppress and OCLC string match. Thanks to Axel, we found that the location filter was not working with the where in the with query above and we moved those to below the main query here. Now the query will provide results only for the location specified.