Update invoice_lines_adjustments.sql
Commented out the code in this derived table to extract the fund distributions data array in the adjustments data array. This data array will be extracted in a separate derived table called invoice_lines_adjustments_fund_distributions in a future release to avoid the possibility of creating duplicate adjustment rows.