Change JSON operations in LDP derived tables to work with both json and jsonb types (#771)
LDP 2.0 is planned to use the `jsonb` data type instead of `json`, which
will be a breaking change. This PR attempts to convert all JSON
operations in LDP derived tables to support both data types, in order to
minimize the impact of upgrading to LDP 2.0.

Note that these changes use the `#>` and `#>>` JSON operators (and other
techniques) to allow us to continue supporting LDP 1.x for some period
of time while also supporting LDP 2.0. However, for Metadb derived
tables, I would still recommend standardizing on functions such as
`jsonb_extract_path_text()` as they are significantly more readable.

No need to approve this PR, but comments are welcome.