RA/UM - Fixes for derived table 'instance_ext' for Metadb (#773)
Closes #772 

The derived table "instance_ext" for metadb did not correspond 1:1 to
the version for LDP1. Ported reports for MetaDB using this table did not
work. The attributes now match those of the LDP1 version. Two attributes
that were only in the Metadb version have been left so as not to
interfere with any existing reports.