Added calculable transaction amounts to finance_invoice_transactions
- Budget calculations may require transaction amounts of the Credit
transaction type to be in negative amounts. Therefore, an additional
column `effective_transaction_amount` was added to the derived table.
This column contains all transaction amounts for calculations.
- Updated the column `transaction_fund_id` to merge the fund IDs of the
column from_fund_id with the fund IDs from the column to_fund_id into
one column. This is required because the fund ID for credits are in the
`to_fund_id` column. In order not to change the column for fund IDs of
the derived table, these fund IDs were always inserted into the existing
column `transaction_fund_id`.